amazon_scraper/scraper.py
# 亚马逊商品爬虫
# 负责搜索匹配商品、获取高清图片和视频

import requests
from bs4 import BeautifulSoup
from playwright.sync_api import sync_playwright
from loguru import logger
from dataclasses import dataclass
from typing import List, Optional
import time

@dataclass
class AmazonProduct:
    """亚马逊商品数据结构"""
    asin: str
    title: str
    price: float
    currency: str
    rating: float
    reviews_count: int
    main_image: str
    images: List[str]
    video_url: Optional[str]
    url: str
    category: str
    is_prime: bool
    bsr_rank: int  # Best Seller Rank

class AmazonScraper:
    """亚马逊商品爬虫"""
    
    def __init__(self, config: dict):
        self.config = config
        self.base_url = "https://www.amazon.com"
        self.search_url = "https://www.amazon.com/s"
        
    def search_products(self, keyword: str, limit: int = 10) -> List[AmazonProduct]:
        """搜索商品"""
        logger.info(f"Searching Amazon for: {keyword}")
        # TODO: 实现搜索逻辑
        # 建议使用Amazon Product API或爬虫
        return []
    
    def get_product_details(self, asin: str) -> AmazonProduct:
        """获取商品详情"""
        logger.info(f"Fetching details for ASIN: {asin}")
        # TODO: 实现详情抓取
        return None
    
    def download_images(self, product: AmazonProduct, save_dir: str) -> List[str]:
        """下载商品图片"""
        logger.info(f"Downloading images for: {product.title}")
        local_paths = []
        for i, img_url in enumerate(product.images):
            # TODO: 实现图片下载
            pass
        return local_paths
    
    def find_similar_products(self, tiktok_product_name: str) -> List[AmazonProduct]:
        """根据TikTok商品名查找相似亚马逊商品"""
        logger.info(f"Finding similar products for: {tiktok_product_name}")
        return self.search_products(tiktok_product_name, limit=5)
    
    def get_product_video(self, asin: str) -> Optional[str]:
        """获取商品视频"""
        # Amazon部分商品有视频展示
        logger.info(f"Fetching video for ASIN: {asin}")
        return None
    
    def check_price_history(self, asin: str) -> dict:
        """检查价格历史（使用Keepa等第三方服务）"""
        # TODO: 集成Keepa API
        return {}
    
    def run(self, keywords: List[str]) -> List[AmazonProduct]:
        """主运行函数"""
        results = []
        for keyword in keywords:
            products = self.search_products(keyword)
            results.extend(products)
            time.sleep(2)  # 避免请求过快
        return results

if __name__ == "__main__":
    scraper = AmazonScraper({})
    products = scraper.search_products("wireless bluetooth earbuds")
    print(f"Found {len(products)} products")
